pyproject.toml
[project]
name = "fedicode-discord-bot"
version = "0.1.0"
description = "Fedicode Discord サーバー向け Bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.0,<3.0",
    "python-dotenv>=1.0",
]
